feat: 修改ui界面，合并条件判断，临时目录生成策略修改

+++ src/combine.rs
@@ -1,6 +1,11 @@
-use std::{env, io::Write, process::Command, string};
+use std::{env, io::Write, process::{Command, Stdio}, string};
 
 pub fn combine_clip(clip_dir: &str, save_path: &str) {
+    let dir_ex = std::fs::read_dir(clip_dir);
+    if dir_ex.is_err(){
+        panic!("clip_dir: {} not exists!", clip_dir);
+    }
+
     let save_path = get_output_name(save_path);
     println!("开始合并片段，cli_dir:{} save_path:{}", clip_dir, save_path);
     // 1. 检测环境变量
@@ -24,6 +29,10 @@ pub fn combine_clip(clip_dir: &str, save_path: &str) {
             let line = format!("file '{}'\n", file_name);
             file_list.push(line);
         }
+        if file_list.is_empty(){
+            println!("合并目录：{}为空", clip_dir);
+            return;
+        }
         file_list.sort_by(|x,y|{
             x.cmp(y)
         });
@@ -39,20 +48,30 @@ pub fn combine_clip(clip_dir: &str, save_path: &str) {
 
     let output_name = save_path;
     // 3.调用合并
-    let output = 
+    let mut child = 
         Command::new(ffmpeg)
                 .arg("-f").arg("concat").arg("-i")
                 .arg(com_file_name.as_str()).arg("-c").arg("copy")
                 .arg(output_name)
-                .output()
+                .stdout(Stdio::inherit())
+                .stderr(Stdio::inherit())
+                .stdin(Stdio::piped())
+                .spawn()
                 .expect("ffmpeg exec error!");
+    let mut stdin = child.stdin.take().expect("stdin take err");
+    std::thread::spawn(move ||{
+        stdin.write_all(b"y\r\n").expect("写入失败");
+    });
     
-    let output_str = String::from_utf8_lossy(&output.stderr);
-    println!("output_str={}", output_str);
-    println!("开始删除临时文件:");
+    let status = child.wait().unwrap();
+    println!("===>output status={}", status);
+    println!("===>output success={}", status.success());
     
-    std::fs::remove_dir_all(clip_dir).expect("删除临时文件失败！");
-    println!("删除临时文件完成！");
+    if status.success() {
+        println!("开始删除临时文件:");
+        std::fs::remove_dir_all(clip_dir).expect("删除临时文件失败！");
+        println!("删除临时文件完成！");
+    }
 }
 
 fn get_output_name(save_path: &str) -> String {
